Move non-upgrading code to a separate DataDependencies module
